backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
google-cloud-aiplatform==1.71.1
vertexai==1.71.1
twilio==9.0.4
requests==2.31.0